requirements.txt
Flask==2.0.2
gunicorn==20.0.4
python-dotenv==0.15.0
requests==2.27.1
